+++ crates/pallets/starknet/src/simulations.rs
@@ -7,7 +7,9 @@ use blockifier::transaction::objects::TransactionExecutionInfo;
 use blockifier::transaction::transaction_execution::Transaction;
 use blockifier::transaction::transactions::{ExecutableTransaction, L1HandlerTransaction};
 use frame_support::storage;
-use mp_simulations::{InternalSubstrateError, SimulationError, SimulationFlags, TransactionSimulationResult};
+use mp_simulations::{
+    InternalSubstrateError, ReExecutionResult, SimulationError, SimulationFlags, TransactionSimulationResult,
+};
 use mp_transactions::execution::{
     commit_transactional_state, execute_l1_handler_transaction, run_non_revertible_transaction,
     run_revertible_transaction, MutRefState, SetArbitraryNonce,
@@ -18,8 +20,6 @@ use starknet_api::transaction::TransactionVersion;
 
 use crate::blockifier_state_adapter::BlockifierStateAdapter;
 use crate::{log, Config, Error, Pallet};
-
-type ReExecutionResult = Result<Vec<(TransactionExecutionInfo, Option<CommitmentStateDiff>)>, SimulationError>;
 
 impl<T: Config> Pallet<T> {
     pub fn estimate_fee(
@@ -267,6 +267,55 @@ impl<T: Config> Pallet<T> {
             .collect::<Result<_, SimulationError>>()?;
 
         Ok(execution_infos)
+    }
+
+    pub fn get_transaction_re_execution_state_diff(
+        transactions_before: Vec<Transaction>,
+        transactions_to_trace: Vec<Transaction>,
+    ) -> Result<Result<CommitmentStateDiff, SimulationError>, InternalSubstrateError> {
+        storage::transactional::with_transaction(|| {
+            let res = Self::get_transaction_re_execution_state_diff_inner(transactions_before, transactions_to_trace);
+            storage::TransactionOutcome::Rollback(Result::<_, DispatchError>::Ok(Ok(res)))
+        })
+        .map_err(|e| {
+            log::error!("Failed to reexecute a tx: {:?}", e);
+            InternalSubstrateError::FailedToCreateATransactionalStorageExecution
+        })?
+    }
+
+    fn get_transaction_re_execution_state_diff_inner(
+        transactions_before: Vec<Transaction>,
+        transactions_to_trace: Vec<Transaction>,
+    ) -> Result<CommitmentStateDiff, SimulationError> {
+        let block_context = Self::get_block_context();
+        let mut state = BlockifierStateAdapter::<T>::default();
+
+        transactions_before.iter().try_for_each(|tx| {
+            Self::execute_transaction(tx, &mut state, &block_context, &SimulationFlags::default()).map_err(|e| {
+                log::error!("Failed to reexecute a tx: {}", e);
+                SimulationError::from(e)
+            })?;
+            Ok::<(), SimulationError>(())
+        })?;
+
+        let mut transactional_state = CachedState::new(MutRefState::new(&mut state), GlobalContractCache::new(1));
+
+        transactions_to_trace.iter().try_for_each(|tx| {
+            Self::execute_transaction(tx, &mut transactional_state, &block_context, &SimulationFlags::default())
+                .map_err(|e| {
+                    log::error!("Failed to reexecute a tx: {}", e);
+                    SimulationError::from(e)
+                })?;
+            Ok::<(), SimulationError>(())
+        })?;
+
+        let state_diff = transactional_state.to_state_diff();
+        commit_transactional_state(transactional_state).map_err(|e| {
+            log::error!("Failed to commit state changes: {:?}", e);
+            SimulationError::from(e)
+        })?;
+
+        Ok(state_diff)
     }
 
     fn execute_transaction<S: State + SetArbitraryNonce>(